--- titiler/api/api_v1/endpoints/ogc.py
@@ -1,90 +1,0 @@
-"""API ogc."""
-
-from urllib.parse import urlencode
-
-from fastapi import APIRouter, Query
-from starlette.requests import Request
-from starlette.responses import Response
-from starlette.templating import Jinja2Templates
-
-from rio_tiler_crs import COGReader
-
-from titiler.core import config
-from titiler.api.deps import TileMatrixSetNames, morecantile
-from titiler.ressources.enums import ImageType, ImageMimeTypes, MimeTypes
-from titiler.ressources.responses import XMLResponse
-
-router = APIRouter()
-templates = Jinja2Templates(directory="titiler/templates")
-
-
-@router.get("/cog/WMTSCapabilities.xml", response_class=XMLResponse, tags=["WMTS"])
-@router.get(
-    "/cog/{TileMatrixSetId}/WMTSCapabilities.xml",
-    response_class=XMLResponse,
-    tags=["WMTS"],
-)
-def wtms(
-    request: Request,
-    response: Response,
-    TileMatrixSetId: TileMatrixSetNames = Query(
-        TileMatrixSetNames.WebMercatorQuad,  # type: ignore
-        description="TileMatrixSet Name (default: 'WebMercatorQuad')",
-    ),
-    url: str = Query(..., description="Cloud Optimized GeoTIFF URL."),
-    tile_format: ImageType = Query(
-        ImageType.png, description="Output image type. Default is png."
-    ),
-    tile_scale: int = Query(
-        1, gt=0, lt=4, description="Tile size scale. 1=256x256, 2=512x512..."
-    ),
-):
-    """Wmts endpoit."""
-    scheme = request.url.scheme
-    host = request.headers["host"]
-    if config.API_VERSION_STR:
-        host += config.API_VERSION_STR
-    endpoint = f"{scheme}://{host}/cog"
-
-    kwargs = dict(request.query_params)
-    kwargs.pop("tile_format", None)
-    kwargs.pop("tile_scale", None)
-    qs = urlencode(list(kwargs.items()))
-
-    tms = morecantile.tms.get(TileMatrixSetId.name)
-    with COGReader(url, tms=tms) as cog:
-        minzoom, maxzoom, bounds = cog.minzoom, cog.maxzoom, cog.bounds
-
-    media_type = ImageMimeTypes[tile_format.value].value
-
-    tileMatrix = []
-    for zoom in range(minzoom, maxzoom + 1):
-        matrix = tms.matrix(zoom)
-        tm = f"""
-                <TileMatrix>
-                    <ows:Identifier>{matrix.identifier}</ows:Identifier>
-                    <ScaleDenominator>{matrix.scaleDenominator}</ScaleDenominator>
-                    <TopLeftCorner>{matrix.topLeftCorner[0]} {matrix.topLeftCorner[1]}</TopLeftCorner>
-                    <TileWidth>{matrix.tileWidth}</TileWidth>
-                    <TileHeight>{matrix.tileHeight}</TileHeight>
-                    <MatrixWidth>{matrix.matrixWidth}</MatrixWidth>
-                    <MatrixHeight>{matrix.matrixHeight}</MatrixHeight>
-                </TileMatrix>"""
-        tileMatrix.append(tm)
-
-    tile_ext = f"@{tile_scale}x.{tile_format.value}"
-    return templates.TemplateResponse(
-        "wmts.xml",
-        {
-            "request": request,
-            "endpoint": endpoint,
-            "bounds": bounds,
-            "tileMatrix": tileMatrix,
-            "tms": tms,
-            "title": "Cloud Optimized GeoTIFF",
-            "query_string": qs,
-            "tile_format": tile_ext,
-            "media_type": media_type,
-        },
-        media_type=MimeTypes.xml.value,
-    )
