--- titiler/api/api_v1/endpoints/operations.py
@@ -1,8 +1,0 @@
-"""API operations."""
-
-# /point
-# /static/{lon},{lat},{zoom}/{width}x{height}.{format}?
-
-# /cog/point
-# /cog/part
-# /cog/preview
